Add support for auto-upgrading dependencies. An example of what this fixes is below. Mod A depends on Version `^0.1.0` of Mod B Mod A is installed, and mod B is installed with that version.
Mod C depends on Version `^0.2.0` of Mod B , and has a download link to this.
Mod C is installed.

In the past, QP would say "Already installed version of B is not within version range requested by C"
Now it will attempt to download the new version, and will look to see if the downloaded version intersects the version range of A, and any other mods that depend on C.
If it does, the existing C will be uninstalled, and the new one installed. (this happens, since A requests `^0.1.0`)
If it doesn't QP will print a message explaining the issue.